requirements.txt
aioice==0.10.1
aiortc==1.13.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
av==14.4.0
backports.functools-lru-cache==2.0.0
beautifulsoup4==4.13.4
bidict==0.23.1
blis==1.3.0
boto3==1.40.14
botocore==1.40.14
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
charset-normalizer==3.4.3
click==8.1.8
cloudpathlib==0.21.1
confection==0.1.5
constantly==23.10.4
cryptography==45.0.6
cymem==2.0.11
daphne==4.2.1
distro==1.9.0
Django==5.2.5
django-channels==0.7.0
dnspython==2.7.0
emoji==2.14.1
filelock==3.19.1
fsspec==2025.7.0
fuzzywuzzy==0.18.0
google-crc32c==1.7.1
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
ifaddr==0.2.0
incremental==24.7.2
indic-transliteration==2.3.74
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
langcodes==3.5.0
language_data==1.3.0
Levenshtein==0.27.1
llvmlite==0.44.0
marisa-trie==1.3.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.5
nltk==3.9.1
numpy==2.2.6

# ---- Linux/GPU specific (COMMENTED OUT for Windows) ----
# nvidia-cublas-cu12==12.8.4.1
# nvidia-cuda-cupti-cu12==12.8.90
# nvidia-cuda-nvrtc-cu12==12.8.93
# nvidia-cuda-runtime-cu12==12.8.90
# nvidia-cufft-cu12==11.3.3.83
# nvidia-cufile-cu12==1.13.1.3
# nvidia-curand-cu12==10.3.9.90
# nvidia-cusolver-cu12==11.7.3.90
# nvidia-cusparse-cu12==12.5.8.93
# nvidia-cusparselt-cu12==0.7.1
# nvidia-nccl-cu12==2.27.3
# nvidia-nvjitlink-cu12==12.8.93
# nvidia-nvtx-cu12==12.8.90
# --------------------------------------------------------

oauthlib==3.3.1
openai==1.100.2
packaging==25.0
preshed==3.0.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.2
pylibsrtp==0.12.0
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-engineio==4.12.2
python-Levenshtein==0.27.1
python-socketio==5.13.0
RapidFuzz==3.13.0
redis==6.4.0
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
roman==5.1
s3transfer==0.13.1
service-identity==24.2.0
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.3
sqlparse==0.5.3
srsly==2.5.1
srt==3.5.3
sympy==1.14.0
textblob==0.19.0
thinc==8.3.6
tiktoken==0.11.0
toml==0.10.2
tqdm==4.67.1
# triton==3.4.0   # mostly Linux + CUDA, comment for Windows
Twisted==25.5.0
txaio==25.6.1
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
vaderSentiment==3.3.2
vosk==0.3.45
wasabi==1.1.3
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.3
wsproto==1.2.0
zope.interface==7.2
